refactor(input box): seperate input widgets in different files
