+++ lib/widgets/input/chip_input_box.dart
@@ -1,30 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:kai_friends_app/widgets/color_chip.dart';
-import 'package:kai_friends_app/widgets/main_button.dart';
-
-class InputLabel extends StatelessWidget {
-  final String _name;
-
-  const InputLabel({
-    required name,
-    Key? key,
-  })  : _name = name,
-        super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: double.infinity,
-      padding: const EdgeInsets.symmetric(horizontal: 12),
-      child: Text(
-        _name,
-        style: TextStyle(
-          color: Colors.grey.shade700,
-        ),
-      ),
-    );
-  }
-}
+import '../color_chip.dart';
+import '../main_button.dart';
+import 'input_box.dart';
+import 'input_label.dart';
 
 class ChipInputBox extends StatelessWidget {
   const ChipInputBox({
@@ -62,7 +40,7 @@ class ChipInputBox extends StatelessWidget {
               // TODO: add multiple ColorChip() to here,
               ActionChip(
                 label: const Icon(Icons.add),
-                onPressed: () => ShowDialog(context),
+                onPressed: () => showDetailDialog(context),
               ),
             ],
           ),
@@ -71,7 +49,7 @@ class ChipInputBox extends StatelessWidget {
     );
   }
 
-  void ShowDialog(BuildContext context) async {
+  void showDetailDialog(BuildContext context) async {
     showDialog(
       context: context,
       builder: (BuildContext context) {
@@ -133,83 +111,6 @@ class ChipInputBoxDetailDialog extends StatelessWidget {
             ],
           ),
         ],
-      ),
-    );
-  }
-}
-
-class DropdownInputBox extends StatelessWidget {
-  const DropdownInputBox({
-    Key? key,
-    required selected,
-    required List<String> list,
-    required labelText,
-  })  : _selected = selected,
-        _list = list,
-        _labelText = labelText,
-        super(key: key);
-
-  final String _selected;
-  final List<String> _list;
-  final String _labelText;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 20),
-      child: DropdownButtonFormField(
-          value: _selected,
-          decoration: InputDecoration(
-            border: const OutlineInputBorder(),
-            labelText: _labelText,
-          ),
-          items: _list.map(
-            (value) {
-              return DropdownMenuItem(
-                child: Text(value),
-                value: value,
-              );
-            },
-          ).toList(),
-          onChanged: (value) {}),
-    );
-  }
-}
-
-class InputBox extends StatelessWidget {
-  const InputBox({
-    Key? key,
-    required labelText,
-    required type,
-    actionIcon,
-    action,
-    paddingAll,
-  })  : _labelText = labelText,
-        _type = type,
-        _actionIcon = actionIcon,
-        _action = action,
-        _paddingAll = paddingAll,
-        super(key: key);
-  final String _labelText;
-  final TextInputType _type;
-  final IconData? _actionIcon;
-  final Function()? _action;
-  final double? _paddingAll;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.symmetric(
-          horizontal: _paddingAll ?? 10, vertical: _paddingAll ?? 20),
-      child: TextField(
-        keyboardType: _type,
-        decoration: InputDecoration(
-          border: const OutlineInputBorder(),
-          labelText: _labelText,
-          suffixIcon: _actionIcon != null
-              ? IconButton(onPressed: _action, icon: Icon(_actionIcon))
-              : null,
-        ),
       ),
     );
   }
